Refactor swagger task
